+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ _Date TBD_
 ### Features Added
 * Add basic search API endpoint to return all certified products
 * Add surveillance statistics to weekly email
+* Protect basic surveillance report and only allow download by ONC ADMIN and ONC STAFF
 
 ---